Add compiled m5stack-cardputer firmware binary
Built with PlatformIO (espressif32 55.3.32, arduino-esp32 3.3.2).
Includes flash backup feature, single-file backup format, filename
prompt, and yellow theme boot screen colour fix.

Also fixes a size_t/uint32_t type mismatch in backupPartitionToFile,
and makes the web-file minifier fall back to raw content when the
Toptal API is unreachable. Adds Launcher-*.bin exception to .gitignore
so release binaries can be tracked.

+++ src/sd_functions.cpp
@@ -674,7 +674,7 @@ static bool backupPartitionToFile(
 
     size_t done = 0;
     while (done < partition->size) {
-        const size_t toRead = min(kBufSize, partition->size - done);
+        const size_t toRead = min(kBufSize, static_cast<size_t>(partition->size) - done);
         if (esp_partition_read(partition, done, buf.get(), toRead) != ESP_OK) return false;
         if (out.write(buf.get(), toRead) != toRead) return false;
         done += toRead;
